Feat(Payment): 토스 결제 및 레디스
- 분산락 적용
- 검증 로직 적용

+++ src/main/java/com/example/pos/dto/pg/PaymentRequestDTO.java
@@ -1,6 +1,7 @@
 package com.example.pos.dto.pg;
 
 import lombok.*;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -9,5 +10,5 @@ import lombok.*;
 public class PaymentRequestDTO {
     private String orderId;
     private String paymentKey;
-    private int amount;
+    private Long amount;
 }
